Fixed repeated invocation of block for GCDWebServerStreamedResponse
When attempting to use the GCDWebServerStreamedResponse as described
in the README, repeatedly invoking the GCDWebServerBodyReaderCompletionBlock
to signal the readiness of more data, a number of bad things happen:

-The read loop in GCDWebServerConnection writeBodyWithCompletionBlock()
repeatedly invokes the body reader completion block attached to the
streamed response object (because of a recursive call to
writeBodyWithCompletionBlock).
I.e. every time the BodyReaderCompletionBlock is called by the
asyncStreamBlock, the asyncStreamBlock is invoked again.

-This leads to a large number of "broken pipe" warnings in the console
on iOS8, at least. On a larger project it with many requests it would
reliably result in an EXC_BAD_ACCESS, but I have not been able to reproduce
this on a small project.

This problem is fundamentially due to the differences in the sync and async
body model. For synchronous responses, they are expected to return some
data from performDataReadWithCompletion:, and the connection repeatedly
calls this method until there is no data left.

For asynchronous responses, the user of the library can happily invoke
the BodyReaderCompletionBlock multiple times until they are done; there is
no need for the connection to have a loop driving this process.

This commit addresses this issue by only repeatedly calling
performDataReadWithCompletion: for synchronous response types