Iterate over a copy of _handlers to prevent a crash
Since we're modifying the handlers when the server is running, it may
sometimes crash when processing a connection. This change prevents the
crash by first creating a copy of the array to iterate over.

Respective crash example:
https://fabric.io/8fit2/ios/apps/com.8fit.app/issues/5724f4bfffcdc042503633aa

Possibly, it fixes also
https://fabric.io/8fit2/ios/apps/com.8fit.app/issues/5734468affcdc04250b72a30
because there's no other place where `handlers` is modified.